CPR-264 Add a search service to find candidates
